Add support for ActionFrontControllerSetVariablesController
This update introduces support for the ActionFrontControllerSetVariablesController within the ReCaptcha Module. Additionally, the update introduces checks for whether module is active and if settings are configured, a hook for ActionFrontControllerSetVariables, and it detects if PrestaShop version is 8 or not. Various new functions have been added and existing ones updated to facilitate these new features.