Update Installer hooks and module tab setting
The commit modifies the Installer class in the drsoftfrrecaptcha module. It now defines an array of HOOKS with 'actionFrontControllerSetVariables'. Also, a change has been made in the drsoftfrrecaptcha.php where the tab property of the module was updated from 'emailing' to 'front_office_features'. These changes are meant to enhance the functionality of the reCaptcha feature in the front office of the application.