Refactor reCAPTCHA event listener and form handling
Removed option for null parameter in 'formElm' where the form is added to the reCAPTCHA attributes and event listener. The 'handleSubmit' function now only receives the form element to be submitted, moving the button 'btnElm' and null handling for 'formElm' outside. This results in a more straightforward flow where required elements are checked before adding the reCAPTCHA and the respective event listener.